+++ src/main/java/com/emis/EMIS/repositories/GuardianRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface GuardianRepo extends JpaRepository<GuardianEntity,Integer> {
+public interface GuardianRepo extends JpaRepository<GuardianEntity,Long> {
     List<GuardianEntity> findByStatus(Status status);
 
     GuardianEntity findByGuardianId(int guardianId);
